bug fixes + make subject and body editable in staging
